# 07_API_SPECIFICATION.yaml
openapi: 3.0.3
info:
  title: OIL SIF Precursor Detection API
  description: >
    API for ingesting safety reports, retrieving SIF classifications and
    Life-Saving Rule tags, exploring precursor clusters, and capturing
    analyst feedback.
  version: 0.1.0
servers:
  - url: https://api.internal.oil-sif.example/v1
    description: Prototype/pilot environment
...
paths:
...
  /dashboard/sif-density:
    get:
      tags: [Dashboard]
      summary: SIF-precursor density ranking by site/activity
      security: [{ bearerAuth: [] }]
      parameters:
        - name: date_from
          in: query
          schema: { type: string, format: date }
        - name: date_to
          in: query
          schema: { type: string, format: date }
        - name: group_by
          in: query
          schema: { type: string, enum: [site, department, activity], default: site }
      responses:
        '200':
          description: Ranked density data
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    group_label: { type: string }
                    sif_count: { type: integer }
                    total_count: { type: integer }
                    sif_rate: { type: number, format: float }
  /dashboard/lsr-distribution:
    get:
      tags: [Dashboard]
      summary: Distribution of Life-Saving Rule tags
      security: [{ bearerAuth: [] }]
      parameters:
        - name: site_id
          in: query
          schema: { type: string, format: uuid }
        - name: date_from
          in: query
          schema: { type: string, format: date }
        - name: date_to
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: LSR frequency breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    lsr_category: { type: string }
                    count: { type: integer }
  /dashboard/trend:
    get:
      tags: [Dashboard]
      summary: SIF-potential rate vs total report volume over time
      security: [{ bearerAuth: [] }]
      parameters:
        - name: site_id
          in: query
          schema: { type: string, format: uuid }
        - name: interval
          in: query
          schema: { type: string, enum: [day, week, month], default: week }
      responses:
        '200':
          description: Time series data
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    period: { type: string, format: date }
                    sif_count: { type: integer }
                    total_count: { type: integer }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...